changed: move VideoInfoDownloader to video folder
