+++ xbmc/video/VideoInfoDownloader.h
