+++ xbmc/video/windows/GUIWindowVideoBase.cpp
@@ -20,14 +20,14 @@
 
 #include "video/windows/GUIWindowVideoBase.h"
 #include "Util.h"
-#include "utils/VideoInfoDownloader.h"
 #include "utils/RegExp.h"
 #include "GUIInfoManager.h"
 #include "addons/AddonManager.h"
 #include "addons/IAddon.h"
 #include "utils/GroupUtils.h"
+#include "video/VideoInfoDownloader.h"
 #include "video/dialogs/GUIDialogVideoInfo.h"
-#include "video/windows/GUIWindowVideoNav.h" 
+#include "video/windows/GUIWindowVideoNav.h"
 #include "dialogs/GUIDialogFileBrowser.h"
 #include "video/dialogs/GUIDialogVideoScan.h"
 #include "dialogs/GUIDialogSmartPlaylistEditor.h"
